requirement.txt
mtcnn
face_recognition
opencv-python
pillow
mysql-connector-python
pandas 
numpy 
opencv-contrib-python 
tkcalendar
bcrypt